test for profile exist before append

+++ install.ps1
@@ -1,4 +1,9 @@
-[string[]]$output = "`n`$functions_dir = `"$PSScriptRoot`""
-$output += "Get-ChildItem -Path `$functions_dir -File -Filter *ps1  | %{ Import-Module `$_.FullName -ErrorAction SilentlyContinue}"
-$output |
-    Out-File -FilePath $PROFILE -Append -NoClobber -Encoding ascii
+[string[]]$output = "`$functions_dir = `"$PSScriptRoot`""
+$output += "Get-ChildItem -Path `$functions_dir -File -Filter *ps1  | %{ Import-Module `$_.FullName }"
+if (Test-Path $PROFILE -PathType Leaf) {
+    $output |
+        Out-File -FilePath $PROFILE -Append -NoClobber -Encoding ascii 
+} else {
+    $output |
+        Out-File -FilePath $PROFILE -Encoding ascii
+}
